Disables pylint W0223 in VMImageHtmlParser
The W0223 pylint check warns about any abstract method which is not
implemented by subclasses. Currently this warning are being raised for
VMImageHtmlParser (Python 3.7). However it seems it happens because an
open Python bug[1] about HtmlParser not implementing the error()
abstract method of (inner) superclass _markupase.ParserBase().

To enable avocado project checks for cases like that, this change
disables the related check only for the specific case.

[1] - https://bugs.python.org/issue31844

Signed-off-by: Caio Carrara <ccarrara@redhat.com>