Makefile: improve clean up of plugins
The build directory of the main Avocado module is already cleaned by
the `clean` rule, but when plugins contain a build directory, they are
not currently cleaned.

Sometimes this leads to annoying `inspekt` failures, because previous
different versions of the plugin may be in the build directory, and
the error message only mentions the name of the module/class/method
violating the check, but not the exact file name.

Signed-off-by: Cleber Rosa <crosa@redhat.com>