Extend the LV/VG list functions with optional volume group argument
In addition, we unify the naming of both to standard "vg/lv_list". It
is true that the "--all" switch has different meaning for each of these
functions but we simply default to listing as much as possible in a
single list function for each of LV/VG (e.g. the "--all" switch includes
internal logical volumes the listing of which won't harm but help the
output to be as explicit as possible).

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>