External Runner: support relative paths given on the command line
I was trying to use the external runner, in a reasonably simple way:
from within a project (QEMU) build dir, I'd run:

 $ avocado run --external-runner=tests/qemu-iotests/check \
   --external-runner-chdir=runner -- 001 002

But, what should work, did not.  The problem is with the non-absolute
path to the external runner, that is needed because we're changing
directories.

Absolute paths won't hurt otherwise (when no chdir is done), so it's
always done here.

Signed-off-by: Cleber Rosa <crosa@redhat.com>