scripts/avocado: use a more standard behavior on import failure
I tend to favor more predictable and standard behavior, and the
currently used approach to "protect" the import of Python standard
libraries seems to me to be far from standard.

With this change, if the system's Python interpreter is broken, or if
Python standard libraries can not be imported/loaded, the default
behavior will be executed.  If it seems to be an Avocado problem, then
we try to print a line describing that.

Signed-off-by: Cleber Rosa <crosa@redhat.com>