YAML to Mux: unwind the handle_control_path utility function
The function where this utility method lives is really big and
complex.  Let's break it down for the sake of code readability,
while adding better docstrings.

I've timed the performance changes caused by this, and on most
trials they presented similar performance.

This also adds new tests for this specific unit of code.

Signed-off-by: Cleber Rosa <crosa@redhat.com>