YAML to Mux: unwind the normalize_path utility function
The function where this utility method lives is really big and
complex.  Let's break it down for the sake of code readability,
while adding better docstrings.  At the same time, this allows
us to write tests for those smaller units.

I've timed the performance changes caused by this, and on all
trials they presented slightly better (but overall insignifcant)
improvements.  This is just mentioned to assert that there are
no performance downsides to this change.

Signed-off-by: Cleber Rosa <crosa@redhat.com>