avocado.utils.cloudinit: support root login
If the username is "root", let's also allow it to be used for network
based logins such as SSH.  When this is set to True (the default),
any attempt to connect via SSH will result in a message similar to:

"Please login as user root rather than user root".

Which only makes sense for when another user is configured.

Signed-off-by: Cleber Rosa <crosa@redhat.com>