vmimage: more tests for parser and base and minor improvements
This change focus mostly on increasing tests for VmImageHtmlParser and
ImageProviderBase. During the tests implementation some improvements
were detected: the process of feed a html parser run both for get an
image version and image url was duplicated so it was unified in a single
method.

Besides that the process of get a best version for an image was put on a
specific method so each provider can override it if required. Today the
default rule for a best version is based on calling max() function on
found versions. However we can't assume it's the default for every
provider. As far as we know, for a possible OpenSuse provider in future
this rule will not be valid.

Reference: https://trello.com/c/vtM1SPpJ
Signed-off-by: Caio Carrara <ccarrara@redhat.com>