Make package query uniform across distro rhel and ubuntu
Currently RHEL provides function accepts just filename, absolute
path of file is not required to be passed. But provides function
for Ubuntu expects users to send the absolute path. This commit
makes changes to Ubuntu provides code to make it generic. So
the consumers just can call provides function with filename.

Signed-off-by: Srikanth Aithal <sraithal@linux.vnet.ibm.com>