Add tests for kernel download url building
The current utils kernel module has no tests. Recently it was added a
fix[1] to the process of kernel url building. To keep this process
without future regressions, this change adds a proper test for it.

It's important to note that a new internal method was added to the
KernelBuild class to make easier to test it. Indirectly it's an
improvement of download method that now is shorter and more explicit.

[1] - commit c84ac6ba5eb892a54128e42649b9beb53fc1c20e

Signed-off-by: Caio Carrara <ccarrara@redhat.com>