Job: internal variable rename
The related command line option is "--keep-tmp", and it keeps the
setting on "keep_tmp", so it just makes sense, for easier reading of
the code, to name the variable in a similar way.

While at it, let's prevent crash if "os.path.exists()" in "cleanup()"
is given a tmpdir that is None.  This was not caught before because
the default value of "__remove_tmp" would prevent "os.path.exists()"
to be reached.

Signed-off-by: Cleber Rosa <crosa@redhat.com>