selftests/unit/test_utils_partition.py: simplify mount usage
The goal of the shell command used in these tests is to simply have
the mountpoint as being used, and for that, there's no need to keep
creating/removing files (it's actually an invitation to race
conditions).

Let's replace that approach with a more stable/predictable command,
which opens a single file for writing at the mount point location.

The original motivation for the change was caused by failures on
the test when run in certain environments.  I believe the fact
that trying to kill the shell process while it's spawning "echo"
and "rm" in a busy loop could be causing a race condition.

Signed-off-by: Cleber Rosa <crosa@redhat.com>