selftests/unit/test_utils_partition.py: add a real force unmount
Because the test test_force_unmount_get_pids_fail() mocks the
inner behavior of the the unmount() method, let's add a call
to non-mocked version before the process attempts to terminate
itself.

Note: this is a hot fix for the problem described in the comment,
and referenced below.

Reference: https://trello.com/c/7OnKaLoG/1446-bug-procterminate-not-reliable
Signed-off-by: Cleber Rosa <crosa@redhat.com>