avocado/utils/lv_utils.py: preserve the original parameter order of lv_create()
To minimize brekage of existing test code.

Even though the API may looks better on the last commit, it's better
to add the new parameters to the end. That is, force_flag should
continue being the third parameter, so users using positional
parameters won't be surprised by the order change.

Signed-off-by: Cleber Rosa <crosa@redhat.com>