selftests/unit/test_utils_partition.py: increase timeout to create file
Another process is spawned in this test, and it's supposed to create a
file within the mountpoint.  It looks like the 1s timeout is not
enough on very low powered environments.

At the same time, let's add an explicit write and sync to ask the I/O
scheduler to write the file to disk.

Signed-off-by: Cleber Rosa <crosa@redhat.com>