ISO9660 selftests: use mnt_dir_workflow on the base class
The division of responsiblities in the ISO9660 test classes is one that,
because the children classes need to setup the "iso" instance, the
"mnt_dir_worflow" is not named as a method that is considered a test
in the base class.

The problems with that are:

 1) a lot of boiler plate code in the children classes just to call
    the "mnt_dir_workflow" method.
 2) skip*() decorators do not work in generic methods.

To simplify issue 1, and fix 2, let's rename that method to become
a test method, and just add a condition that, unless "self.iso" is
prepared, the test is going to be skipped.

Signed-off-by: Cleber Rosa <crosa@redhat.com>