Software Manager: do not ignore apt-file status to record log
The try/except block around the execution of apt-file to update
is pointless if we're ignoring the status of the apt-file command.

I'll defer to a later moment the discussion on whether this failure
should be fatal, and not just log its failure.  That is currently
covered by the fact that the "apt-file search <path>" execution
will fail with non-zero status when the cache is empty (that is,
when apt-file update was not executed successfully).

Also, add a test that checks the behavior of the affected method.

Signed-off-by: Cleber Rosa <crosa@redhat.com>