Adds rate property to Result class
Currently to generate the HTML report the success rate of a test result
is computed by the plugin itself. Since this is an information about the
test result it should be calculated by the Result object and no anywhere
else.

This change adds the rate property to Result class according with the
code (method) already in place in HTML plugin. This new property will be
used soon when the new template engine be added to HTML plugin.

Signed-off-by: Caio Carrara <ccarrara@redhat.com>