avocado.utils.path: introduce check_readable()
It's a common use case to want to make sure a file can be read, or if
not fail as early as possible with a relevant error message.

The avocado.utils.iso9600 is doing that itself, but for the sake of
being useful to a wider audience, let's make that a utility itself.

As always, the benefits are also increased test coverage and
documentation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>