utils.process: Avoid waiting for stopped process when timeout set
Previous recursive implementation of waiting for children is not really
possible, because parent has to treat the children exit, otherwise we
get temporary zombie status until the parent process is resumed.

This implementation also uses recursive way of killing the children, but
instead of waiting for them to die it reports the killed processes and
only in the main function it resumes all of them and watches them to
die.

As a benefit people might now get list of killed processes as return.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>