HTML plugin: always escape content when rendering the report
We only render HTML, that is, we never render anything that must not
have any content escaped. So, let's always enable the autoescape
feature.

This fixes the issue with jinja2 version 2.7.2, available on EL 7,
that has the autoescape parameter to environments, but doesn't have
the select_autoescape() utility function for selecting what to escape
automatically.

Signed-off-by: Cleber Rosa <crosa@redhat.com>