Tags: replace internal representation from set to dict
This basically allows us to use the same set interface for checking if
a tag is in a set (the dict keys), and at the same time, allows us to
have room for the values given to key:val tags.

For "flat" tags, that is, ones that do not contain any value, the
dictionary will just contain None, and for the key:val tags, the
dictionary will contain a set of the values.  The set is useful
because duplicate values add no value, and a common use case
is to have a base value at the class level, and have it extended
in the

Signed-off-by: Cleber Rosa <crosa@redhat.com>