Tags: add support for tag key/val match
Up to this point, the matching and thus filtering of tags was flat.
Now, the individual value of the required keys are checked.

For instance, if you have a test such as:

    class My(Test):
        def test(self):
            """
            :avocado: tags=classification:tier1
            """
            ...

Both of these filter conditions will match:

 $ avocado run -t classification test.py
 $ avocado run -t classification:tier1 test.py

But this won't match:

 $ avocado run -t classification:tier2 test.py

Also, similarly to the how there's support for including tests that
have no tags at all, there's support for including tests that are
not tagged with a specific keys.

Signed-off-by: Cleber Rosa <crosa@redhat.com>