Tags: fix filtering tests without tags and negative matches
The "include_empty" parameter to test the filtering utility function
is meant to act when a test with no tags is part of a test suite that
also includes tests with tags.

There's a current corner case when filtering with tags that must not
be present and tests with no tags, when "include_empty" is active.

Signed-off-by: Cleber Rosa <crosa@redhat.com>