Asset fetcher: handle URL + hash use case
As a documented feature of the asset fetcher, a user may want to make
sure, by the use of a hash, that the fetched asset is sound.  Also, by
giving a name of the asset, allow it to be fetched from a different
number of locations.

But, there's a difference in intention when a name+hash is used, and
when a specific location and hash (a URL+hash) is used.  In the later
case, we don't want to cache the file by name, because different URLs
can have completely different contents.

Signed-off-by: Cleber Rosa <crosa@redhat.com>