utils.process: Avoid hangs
This code is only accessible when timeout is specified so we do want to
finish as soon as possible. Let's replace the "wait" with "poll" which
might lead to declaring process Zombie even when it's not, but it's
already passed 2 timeouts so I think it's better to finish and report
something than to insist on waiting.

This improves a real bug where Avocado hanged after process execution
with timeout. This version is not the full fix, it only declares such
process as zombie and avoids the complete hang.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>