YAML to Mux plugin: use downstream versions in the SPEC file only
Commit c60f0ae85 fixed a CVE warning that GitHub sent notifications
about.  While so, it was attempted to fulfill the downstream (Fedora
and EPEL) dependencies in the same patch, keeping users of Python 2
with possibly vulnerable versions.

The matching of versions that ship in the Fedora and EPEL distros are
really best done in the SPEC file.  Finally, this was motivated by the
fact that the simple Python version check was not enough because
PyYAML>=4.2b2 is not available on Fedora 28, even though it contains
Python 3.

Reference: https://trello.com/c/3UMFOyzQ/1484-pyyaml-requirement-not-fulfilled-on-f28
Signed-off-by: Cleber Rosa <crosa@redhat.com>