selftests/unit/test_utils_vmimage.py: do not depend on host arch
Some tests for the vmimage module do not specify the arch to use to
determine the image that will (would) be fetched.  It then relies on
the default of using the host arch.  But, if the test is run on a host
arch different than x86_64, the test will fail.

Let's make the test more deterministic by providing the arch parameter
value.

Reference: https://trello.com/c/92VQRMoC/1483-vmimage-tests-fail-under-x8664
Signed-off-by: Cleber Rosa <crosa@redhat.com>