avocado.utils.ssh: set the master control path option
The connection sharing functionality of SSH, on which the
avocado.utils.ssh.Session is based upon, requires setting control path
to be given to both the SSH holding the master connection, and to the
SSH instance that should eventually reuse the master connection.

That setting was completely absenting from the library, and the only
reason it worked, was because my local SSH configurations
(~/.ssh/config) had a default setting.

To avoid clashes (even if they could, in theory, be beneficial), let's
flag the control path for the connections initialized by Avocado as
such.

Reference: https://github.com/avocado-framework/avocado/issues/2900
Signed-off-by: Cleber Rosa <crosa@redhat.com>