Release test plan: use RPMs from COPR
The local RPM build is now pretty insignificant, given that we have
RPM builds runnings out of Git, from the master branch on COPR.

Then, if we don't have local RPM files, it's not trivial to run
rpmlint on them.  Let's run it on the host we install the packages
into.

Reference: https://trello.com/c/ihZUtKfc/1347-rpmlint-check-if-which-warnings-errors-can-be-fixed
Reference: https://copr.fedorainfracloud.org/coprs/g/avocado/avocado-latest/
Signed-off-by: Cleber Rosa <crosa@redhat.com>