Release test plan: remove RPM build and lint steps
The local RPM build is now pretty insignificant, given that we have
RPM builds runnings out of Git, from the master branch on COPR.

Then, if we don't have local RPM files, it's not trivial to run
rpmlint on them.  At the same time, the lint checking has never been
effective (warnings can't be ignored), so this is not a major loss.
Instead of linting the RPM, we'll favor functional testing of the RPM
packages on further work.

Reference: https://trello.com/c/ihZUtKfc/1347-rpmlint-check-if-which-warnings-errors-can-be-fixed
Reference: https://copr.fedorainfracloud.org/coprs/g/avocado/avocado-latest/
Signed-off-by: Cleber Rosa <crosa@redhat.com>