Safeloader: return an ordered dictionary
Safeloader is supposed to keep the order of the classes and methods as
they're present on the source code file.  When not using a ordered
dict as an data structure, we count on pure luck.

This addresses the failure of the test_dont_detect_non_avocado test on
unit.test_loader.TestLoader.

Signed-off-by: Cleber Rosa <crosa@redhat.com>