Deployment check: deploy, bootstrap and list Avocado-VT tests
Avocado-VT is a separate project, but it's a very important one,
and it should always be kept working with Avocado.

This adds the Python dependencies that are hard or impossible to
install via PIP, and so that they're available on the virtual
environment, a "site packages" options is set.

We then bootstrap Avocado-VT and finally sure that one of the common
tests is present.

Signed-off-by: Cleber Rosa <crosa@redhat.com>