Selftests: pin pylint and astroid version
The pylint version in use requires a specific version (or a not
so greater version) of astroid.  When the astroid version is not
pinned, astroid version 2.2.4 currently gets installed.

Then, pylint (and thus inspektor) fails with:

  'Import' object has no attribute 'infer_name_module'

Let's bump the pylint version (already pinned) and pin a compatible
astroid version too.

Signed-off-by: Cleber Rosa <crosa@redhat.com>