Docs: add link to LTS version of the Copr repo
The "avocado-latest" Copr repo currently points to packages that are
built with the latest master branch.  Now, there's also a
"avocado-latest-lts" Copr repo, which points to the "latest state of
the latest LTS" branch, which means, the 69lts branch.

So, if a fix makes its way to 69lts, and users want that fix before a
minor version is released, they'd find that fix in the packages
generated from this repo.

Signed-off-by: Cleber Rosa <crosa@redhat.com>