Drop GDB and associated "avocado.utils.runtime" module
This basically removes the transparent command execution plugin, along
with an ill designed mechanism that lets some utility libraries, in
this case `avocado.utils.process`, change its behavior depending on
which test is running, and provide feedback to the currently running
test too.

The correct approach would be for on individual test runner to set the
configuration and enable such utility code for one specific test, and
not the job runner, to configure such things for the "currently
running test", determined by the `runtime.CURRENT_TEST` variable.

This feature may be re-enabled later, once a more capable test runner
mechanism is implemented.  Also note that a second part of the GDB
functionality, the underlying utility libraries that can be used in
tests, are left unchanged.

Signed-off-by: Cleber Rosa <crosa@redhat.com>