selftests/unit/test_utils_cloudinit.py: not require sequential ports
The network utility used to (attempt to) get a free port defaults to
picking sequential ports.  But, that's not a requirement of this test.
When running tests in parallel, the same port may get allocated when a
sequential approach is used, and that is greatly minimized when random
ports are requested.

Signed-off-by: Cleber Rosa <crosa@redhat.com>