Travis-CI: run builds for all branches
When working on a private fork and topic branches, it's helpful to not
have this limitation, and have builds triggered for them.

For the main repo, it should not a problem given that branches are not
often (only on LTS releases), and those are also desirable to have
builds run.

Signed-off-by: Cleber Rosa <crosa@redhat.com>