safeloader: small optmization to avoid parsing avocado tags
On `_examine_class()`, the original assigment of the avocado tags to
the `cl_tags` variables is not necessary because when it's used in the
first occurrence, a `continue` statement follows the assigment, so
it's never reused.  On the other hand, it may be assigned without ever
being used.

On `find_avocado_tests()`, it's just an opmization because `cl_tags`
was only being used once in the function, and at the same time matches
the style used in the previous function.

Signed-off-by: Cleber Rosa <crosa@redhat.com>