Loader: implement an improved Python unittest discovery
The current Python unittest discovery is too permissive, in the sense
that it will match any "test*" method on any class, wether it inherits
from the TestCase class of the unittest module or not.

This new implementation is more strict, but a lot more capable too.
This makes the Python unittest detection very similar (in code and
capabilities) to the Avocado instrumented test detection.

Signed-off-by: Cleber Rosa <crosa@redhat.com>