selftests/unit/test_utils_service.py: simplify the inheritance
A base class exists in this module, and number of the test classes
inherit from it, from the single supposed benefit of sharing a
(static) method.

Let's flatten the inheritance, and make the static method a utility
function, because, if not for anything else, It improves code
readability.

Note: this was actually observed while working on the safeloader
Python unittest detection improvements, meaning that in theory the
safeloader improvements would make this unnecessary for the actual
loading support.  Still, I believe there are benefits (readability,
maintenance) with these changes.

Signed-off-by: Cleber Rosa <crosa@redhat.com>