safeloader: keep track of intermediary module names
When a module with more than one level is imported, such as:

   import unittest.mock

All levels of the module will be available, that is, it's possible to
write code that makes use of whathever is in `unittest`, for instance,
`unittest.TestCase`.

Let's keep track of all intermediary levels imported to improve the
behavior of PythonModule.

Signed-off-by: Cleber Rosa <crosa@redhat.com>