nrunner: add status service and server implementation
One of the task's responsibility is to provide status updates on its
execution.  The actual methods, protocols and channels that a task can
use to send those status updates are not strictly determined.

This initial implementation uses a simple TCP socket based
client/server.  The task execution is currently dependent on sending
the status update, that is, if the given status server is not
available, the task execution will fail.  One of the next planned
features is to allow for a much more reliable execution, independent
of the status server, and one that can store status updates locally
until they can be transmitted (if ever) to a server.

Signed-off-by: Cleber Rosa <crosa@redhat.com>