Makefile: use "avocado nrun" instead of selftests/checkall on check
This switches the custom shell based script that is used to run
Avocado's own tests for the experimental runner.

It brings a native way to run tests in parallel, and an agnostic view
of test types (as far as recognized and supported), as can be seen by
the execution of both shell based tests and Python unittests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>