Travis-CI: check the HEAD of the branch only
For a long time we have resorted to a hack, let's call it "all commits
hack", to test all commits in a branch on Travis-CI.  While it clear
had a positive side, it also had its annoyances.

Now, more and more often, we're faced with the obligation of breaking
PRs in smaller pieces just because of the limitations that the free
tier of Travis-CI imposes (for instance max job time is 50 minutes).

For instance, as this is written, the status of the master branch
is broken, because of the max job time.  And there's no way to
waive it or tweak it.

Also, the hack caused some really interesting problems, and caused
a lot of time wasted, for instance, because it masks the merge commit
that Travis-CI creates, and consequently, makes the signedoff-check.sh
fail on the real HEAD of the merged branch.  Basically, the "all
commits hack" is a requirement to currently have signedoff-check.sh
passing on Travis-CI.

Signed-off-by: Cleber Rosa <crosa@redhat.com>