Varianter CIT Plugin: fix data type of variant nodes
The data type expected to be given to each yield variant contains the
variant id, the "variant" itself, which should be a list of
TreeNode-like instances, and finally the applicable default search
paths.

The "variant" itself, has to be a list, given that on the very first
varianter implementation, it's supposed to contain one item for each
branch of the tree.

Signed-off-by: Cleber Rosa <crosa@redhat.com>