Test: implement multiple setup attempts for INSTRUMENTED tests
This adds a new feature to INSTRUMENTED tests, in which the setUp()
method will attempt to be executed until success is obtained, for a
given number of times.  The default chosen is 3, but that is
configurable.

Given the nature of Avocado INSTRUMENTED tests, which are mostly used
for functional tests, it's understood that setup may involve
interaction with the outside world, which is not always reliable.

For tests that don't have a setUp() that is designed to be executed a
number of tests, the worst that can happen is that the second run will
fail in either the same or in a different way as the first run.

Signed-off-by: Cleber Rosa <crosa@redhat.com>