restclient: drop unmaintained feature
Avocado's restclient tool, and also the avocado-server related tool
avocado-journal-replay, are features that have not received any
maintenance for a long time.  The reason is because avocado-server
itself hasn't been maintained.

One of the post 69.0 goals was to make Avocado slimmer, and try to make
its core features depend on nothing but Python, so let's drop this
feature for now.  If avocado-server resurrects, it can always be
brought back from version control.

Signed-off-by: Cleber Rosa <crosa@redhat.com>